Make a search URL for OnePetro

Create a URL that works in OnePetro website

make_search_url(query = NULL, start = NULL, from_year = NULL,
  peer_reviewed = NULL, published_between = NULL, rows = NULL,
  to_year = NULL, dc_type = NULL, how = "any")

Arguments

query
char any words that will be searched

start
int optional to set the starting paper

from_year
int optional to indicate starting year

peer_reviewed
logical optional, TRUE or FALSE

published_between
logical automatic if from_year or to_year are on

rows
int optional. number of papers to retrieve. max=1000

to_year
int optional to indicate end year

dc_type
char optional to indicate if journal, conference paper

how
char default="any". "all" will match exact words

Examples

# Example 1
url_1 <- make_search_url(query = "flowing gradient survey", how = "all")
onepetro_page_to_dataframe(url_1)

#> # A tibble: 10 x 6
#>                                                                          title_data
#>                                                                               <chr>
#>  1                         Detection and Removal of Naphthenate Deposit in Oil Well
#>  2                                Samarang Downhole Integrity Diagnostic Techniques
#>  3                        Adding Value to Gas Lift Evaluation With Minilogger Tools
#>  4                                 Proper Selection of Multiphase Flow Correlations
#>  5                           Estimating Reservoir Pressure From Early Flowback Data
#>  6            Development of an Integrated Production Network Model for Heera Field
#>  7                           Nonstatic Pressure History Analyses for Gas Reservoirs
#>  8            Acidizing: A Small Change in Recipe, A Big Improvement in Production.
#>  9 Influence of Downhole Fluid Behavior Changes on Slickline Operations in Burgan a
#> 10 Field Trial of an Acoustic Downhole Monitoring System Retrofitted in an Existing
#> # ... with 5 more variables: paper_id <chr>, source <chr>, type <chr>,
#> #   year <int>, author1_data <chr>

# Example 2
url_2 <- make_search_url(query = "static  gradient survey", how = "all")
onepetro_page_to_dataframe(url_2)

#> # A tibble: 10 x 6
#>                                                                          title_data
#>                                                                               <chr>
#>  1 Field Trial of an Acoustic Downhole Monitoring System Retrofitted in an Existing
#>  2 Multilayer Flowmeter Testing Combined With 3D Field Modeling To Enhance the Amel
#>  3                                      Insert String Technique Reviving Idle Wells
#>  4                         Multirate Welltest Design - Niger Delta Field Experience
#>  5 Slimhole Drilling With Expandable Sand Screens for Low-Cost, High-Deliverability
#>  6                                      Aquifer Influx Modelling for Gas Reservoirs
#>  7                         Production Monitoring And Optimization On A Mature Field
#>  8 A Case History: The Installation of a System to Restore Full Safety Valve Functi
#>  9               An Investigation on Factors Influencing In-Situ Emulsion Stability
#> 10 Estimation of Static Bottom Hole Pressure from Well-Head Shut-in Pressure for a 
#> # ... with 5 more variables: paper_id <chr>, source <chr>, type <chr>,
#> #   year <int>, author1_data <chr>

# Example 3
url_3 <- make_search_url(query = "downhole flowrate measurement",
      how = "all", from_year = 1982, to_year = 2017)
onepetro_page_to_dataframe(url_3)

#> # A tibble: 7 x 6
#>                                                                         title_data
#>                                                                              <chr>
#> 1 Permanent Real-Time Downhole Flowrate Measurements in Multilateral Wells Improve
#> 2 Study on the Application of Whole Process Monitoring and Control Technology for 
#> 3          Pressure Transient Analysis and Inflow Performance for Horizontal Wells
#> 4          Derivative of Pressure: Application to Bounded Reservoir Interpretation
#> 5                                   Well Test Interpretation In Faulted Reservoirs
#> 6         Achievements of Smart Well Operations: Completion Case Studies for Hydro
#> 7       Two-Phase Transient Testing of Water Injection Wells in Layered Reservoirs
#> # ... with 5 more variables: paper_id <chr>, source <chr>, type <chr>,
#> #   year <int>, author1_data <chr>